+++ weather_station/protocols/protocol_items.h
@@ -16,6 +16,7 @@
 #include "acurite_592txr.h"
 #include "acurite_5n1.h"
 #include "ambient_weather.h"
+#include "solight_te44.h"
 #include "auriol_hg0601a.h"
 #include "oregon_v1.h"
 #include "tx_8300.h"
@@ -23,6 +24,8 @@
 #include "auriol_ahfl.h"
 #include "kedsum_th.h"
 #include "emos_e601x.h"
+#include "bresser_3ch.h"
 #include "vauno_en8822c.h"
+#include "bl999.h"
 
 extern const SubGhzProtocolRegistry weather_station_protocol_registry;